backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
sqlmodel
psycopg2-binary
python-multipart
twilio
GeoAlchemy2
openrouteservice

# Fijamos las versiones a unas estables y compatibles
passlib==1.7.4
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
